+++ crates/radix-engine-toolkit/tests/transaction_types.rs
@@ -15,9 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use radix_engine::system::bootstrap::*;
-use radix_engine::transaction::*;
-use radix_engine::vm::*;
 use radix_engine_interface::blueprints::account::*;
 use radix_engine_interface::blueprints::consensus_manager::*;
 use radix_engine_interface::blueprints::pool::*;
@@ -27,6 +24,9 @@ use radix_engine_queries::typed_substate_layout::two_resource_pool::*;
 use radix_engine_toolkit::transaction_types::*;
 use scrypto_unit::*;
 use transaction::prelude::*;
+
+mod test_runner_extension;
+use test_runner_extension::TestRunnerEDExt;
 
 macro_rules! assert_eq_three {
     (
@@ -2349,103 +2349,4 @@ fn create_pools(
             *multi_resource_pool_pool_unit,
         ],
     )
-}
-
-#[extend::ext]
-impl<E, D> TestRunner<E, D>
-where
-    E: NativeVmExtension,
-    D: TestDatabase,
-{
-    fn preview(
-        &mut self,
-        manifest: TransactionManifestV1,
-    ) -> TransactionReceiptV1 {
-        self.preview_manifest(
-            manifest,
-            vec![],
-            0,
-            PreviewFlags {
-                use_free_credit: true,
-                assume_all_signature_proofs: true,
-                skip_epoch_check: true,
-            },
-        )
-    }
-
-    fn summarize(
-        &mut self,
-        manifest: TransactionManifestV1,
-    ) -> (ManifestSummary, ExecutionSummary) {
-        let receipt = TestRunnerEDExt::preview(self, manifest.clone());
-        if !receipt.is_commit_success() {
-            panic!("Not commit success: {receipt:?}")
-        }
-
-        let manifest_summary =
-            radix_engine_toolkit::transaction_types::summary(&manifest);
-        let execution_summary =
-            radix_engine_toolkit::transaction_types::execution_summary(
-                &manifest, &receipt,
-            )
-            .unwrap();
-
-        (manifest_summary, execution_summary)
-    }
-
-    fn new_validator(
-        &mut self,
-        pub_key: Secp256k1PublicKey,
-        account: ComponentAddress,
-    ) -> (ComponentAddress, ResourceAddress, ResourceAddress) {
-        let manifest = ManifestBuilder::new()
-            .lock_fee_from_faucet()
-            .get_free_xrd_from_faucet()
-            .take_from_worktop(
-                XRD,
-                *DEFAULT_VALIDATOR_XRD_COST,
-                "xrd_creation_fee",
-            )
-            .create_validator(pub_key, Decimal::ONE, "xrd_creation_fee")
-            .try_deposit_entire_worktop_or_abort(account, None)
-            .build();
-        let receipt = self.execute_manifest(manifest, vec![]);
-        let commit_result = receipt.expect_commit_success();
-
-        let address = commit_result.new_component_addresses()[0];
-        let claim_nft = commit_result.new_resource_addresses()[0];
-        let lsu = commit_result.new_resource_addresses()[1];
-
-        self.execute_manifest(
-            ManifestBuilder::new()
-                .lock_fee_from_faucet()
-                .withdraw_from_account(account, VALIDATOR_OWNER_BADGE, 1)
-                .take_from_worktop(VALIDATOR_OWNER_BADGE, 1, "badge")
-                .create_proof_from_bucket_of_all("badge", "proof")
-                .push_to_auth_zone("proof")
-                .register_validator(address)
-                .call_method(
-                    address,
-                    VALIDATOR_UPDATE_ACCEPT_DELEGATED_STAKE_IDENT,
-                    ValidatorUpdateAcceptDelegatedStakeInput {
-                        accept_delegated_stake: true,
-                    },
-                )
-                .drop_auth_zone_proofs()
-                .return_to_worktop("badge")
-                .try_deposit_entire_worktop_or_abort(account, None)
-                .build(),
-            vec![NonFungibleGlobalId::from_public_key(&pub_key)],
-        )
-        .expect_commit_success();
-
-        self.advance_epoch(10);
-
-        (address, lsu, claim_nft)
-    }
-
-    fn advance_epoch(&mut self, by: u64) {
-        let current_epoch = self.get_current_epoch();
-        self.set_current_epoch(current_epoch.after(by).unwrap());
-    }
 }
